+++ public/app/features/dashboard-scene/pages/DashboardScenePageStateManager.ts
@@ -1,4 +1,4 @@
-import { dateMath, getDefaultTimeRange, getTimeZone, locationUtil, UrlQueryMap } from '@grafana/data';
+import { locationUtil, UrlQueryMap } from '@grafana/data';
 import { t } from '@grafana/i18n';
 import { config, getBackendSrv, isFetchError, locationService } from '@grafana/runtime';
 import { sceneGraph } from '@grafana/scenes';
@@ -36,7 +36,7 @@ import { transformSaveModelSchemaV2ToScene } from '../serialization/transformSav
 import { transformSaveModelToScene } from '../serialization/transformSaveModelToScene';
 import { restoreDashboardStateFromLocalStorage } from '../utils/dashboardSessionState';
 
-import { updateNavModel } from './utils';
+import { processQueryParamsForDashboardLoad, updateNavModel } from './utils';
 
 export interface LoadError {
   status?: number;
@@ -905,50 +905,4 @@ export function getDashboardScenePageStateManager(v?: 'v1' | 'v2') {
   }
 
   return managers.unified;
-}
-
-/**
- * Processes query parameters for dashboard loading, normalizing time range and filtering allowed parameters
- */
-function processQueryParamsForDashboardLoad(): UrlQueryMap {
-  const queryParams = locationService.getSearch();
-  const queryParamsObject = locationService.getSearchObject();
-
-  queryParamsObject.timezone = queryParams.get('timezone') ?? undefined;
-  const now = Date.now();
-  const timeZone = getTimeZone({
-    timeZone: queryParams.get('timezone') ?? undefined,
-  });
-  const defaultTimeRange = getDefaultTimeRange();
-
-  const fromQP = queryParams.get('from');
-  const toQP = queryParams.get('to');
-
-  const to = toQP
-    ? dateMath.toDateTime(toQP, {
-        roundUp: true,
-        timezone: timeZone,
-        now: now,
-      })
-    : undefined;
-
-  const from = fromQP
-    ? dateMath.toDateTime(fromQP, {
-        roundUp: false,
-        timezone: timeZone,
-        now: now,
-      })
-    : undefined;
-
-  queryParamsObject.from = from?.toISOString() ?? defaultTimeRange.from.toISOString();
-  queryParamsObject.to = to?.toISOString() ?? defaultTimeRange.to.toISOString();
-
-  // Remove all properties that are not from, to, scopes, version or start with var-
-  Object.keys(queryParamsObject).forEach((key) => {
-    if (key !== 'from' && key !== 'to' && key !== 'scopes' && key !== 'version' && !key.startsWith('var-')) {
-      delete queryParamsObject[key];
-    }
-  });
-
-  return queryParamsObject;
 }
